# vue/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:9998","description":"Generated server url"}],"paths":{"/upload/file":{"post":{"tags":["上传服务"],"summary":"文件上传","operationId":"upload","parameters":[{"name":"file","in":"query","required":true,"schema":{"type":"string","format":"binary"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/api/user/register":{"post":{"tags":["用户服务"],"summary":"注册","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserVO"}}}}}}},"/api/user/login":{"post":{"tags":["用户服务"],"summary":"登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserVO"}}}}}}},"/api/songs":{"post":{"tags":["歌曲服务"],"summary":"（条件）分页查询歌曲","operationId":"page","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageQueryDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultPageVOSong"}}}}}}},"/api/rating":{"post":{"tags":["评分服务"],"summary":"创建/修改评分","operationId":"rate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RatingDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/collect":{"post":{"tags":["收藏服务"],"summary":"收藏","operationId":"collect","parameters":[{"name":"collectionDTO","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CollectionDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/albums":{"post":{"tags":["专辑服务"],"summary":"（条件）分页查询专辑","operationId":"page_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageQueryDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultPageVOAlbum"}}}}}}},"/upload/album/{id}":{"get":{"tags":["上传服务"],"summary":"专辑上传","operationId":"uploadAlbum","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/song/{id}":{"get":{"tags":["歌曲服务"],"summary":"查询歌曲详情","operationId":"info","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultSongInfoVO"}}}}}}},"/api/rating/{id}":{"get":{"tags":["评分服务"],"summary":"查询评分","operationId":"select","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultInteger"}}}}}}},"/api/collect/{id}":{"get":{"tags":["收藏服务"],"summary":"查询收藏状态","operationId":"select_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultBoolean"}}}}}},"delete":{"tags":["收藏服务"],"summary":"取消收藏","operationId":"cancel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/album/{id}":{"get":{"tags":["专辑服务"],"summary":"查询专辑详情","operationId":"info_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultAlbumInfoVO"}}}}}}}},"components":{"schemas":{"ResultString":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"type":"string"}},"title":"数据响应规范结构"},"UserRegisterDTO":{"type":"object","description":"用户注册DTO","properties":{"username":{"type":"string","description":"用户名","maxLength":16,"minLength":6,"pattern":"^[a-zA-Z0-9_\\u4e00-\\u9fa5]+$"},"phone":{"type":"string","description":"手机号（选填，与邮箱至少填一个）","pattern":"^1[3-9]\\d{9}$"},"email":{"type":"string","format":"email","description":"邮箱（选填，与手机号至少填一个）"},"password":{"type":"string","description":"密码","maxLength":20,"minLength":8,"pattern":"^(?=.*[A-Za-z])(?=.*\\d).+$"},"confirmPassword":{"type":"string","description":"确认密码（需与密码一致）","minLength":1}},"required":["confirmPassword","password","username"],"title":"UserRegisterDTO"},"ResultUserVO":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"$ref":"#/components/schemas/UserVO"}},"title":"数据响应规范结构"},"UserVO":{"type":"object","description":"用户注册/登录VO","properties":{"id":{"type":"integer","format":"int64","description":"用户Id"},"token":{"type":"string","description":"Jwt令牌"}},"title":"UserVO"},"UserLoginDTO":{"type":"object","description":"用户登录DTO","properties":{"account":{"type":"string","description":"登录账号（用户名/手机号/邮箱）","minLength":1},"password":{"type":"string","description":"密码","maxLength":20,"minLength":8,"pattern":"^(?=.*[A-Za-z])(?=.*\\d).+$"}},"required":["account","password"],"title":"UserLoginDTO"},"PageQueryDTO":{"type":"object","description":"分页查询DTO","properties":{"sortType":{"type":"string","description":"排序方式","enum":["DEFAULT","AVG_SCORE","COLLECT_COUNT","COMMENT_COUNT","RELEASE_TIME","CREATE_TIME"]},"page":{"type":"integer","format":"int32","description":"页数"},"size":{"type":"integer","format":"int32","description":"记录数"}},"required":["page","size"],"title":"PageQueryDTO"},"PageVOSong":{"type":"object","description":"分页查询VO","properties":{"total":{"type":"integer","format":"int64","description":"总记录数"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/Song"}}},"title":"PageVO"},"ResultPageVOSong":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"$ref":"#/components/schemas/PageVOSong"}},"title":"数据响应规范结构"},"Song":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"neteaseId":{"type":"integer","format":"int64"},"songName":{"type":"string"},"albumId":{"type":"integer","format":"int64"},"albumName":{"type":"string"},"artists":{"type":"string"},"releaseTime":{"type":"string","format":"date"},"duration":{"type":"integer","format":"int32"},"coverUrl":{"type":"string"},"avgScore":{"type":"integer","format":"int32"},"ratingCount":{"type":"integer","format":"int32"},"collectCount":{"type":"integer","format":"int32"},"commentCount":{"type":"integer","format":"int32"}}},"RatingDTO":{"type":"object","description":"评分DTO","properties":{"targetId":{"type":"integer","format":"int64","description":"评分对象id"},"targetType":{"type":"integer","format":"int32","description":"评分对象类型"},"score":{"type":"integer","format":"int32","description":"评分分数"}},"required":["score","targetId","targetType"],"title":"RatingDTO"},"ResultVoid":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{}},"title":"数据响应规范结构"},"CollectionDTO":{"type":"object","description":"收藏DTO","properties":{"targetId":{"type":"integer","format":"int64","description":"收藏对象id"},"targetType":{"type":"integer","format":"int32","description":"收藏资源类型：1-专辑 2-歌曲"}},"title":"CollectionDTO"},"Album":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"neteaseId":{"type":"integer","format":"int64"},"albumName":{"type":"string"},"artists":{"type":"string"},"releaseTime":{"type":"string","format":"date"},"coverUrl":{"type":"string"},"description":{"type":"string"},"avgScore":{"type":"number"},"ratingCount":{"type":"integer","format":"int32"},"collectCount":{"type":"integer","format":"int32"},"commentCount":{"type":"integer","format":"int32"}}},"PageVOAlbum":{"type":"object","description":"分页查询VO","properties":{"total":{"type":"integer","format":"int64","description":"总记录数"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/Album"}}},"title":"PageVO"},"ResultPageVOAlbum":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"$ref":"#/components/schemas/PageVOAlbum"}},"title":"数据响应规范结构"},"ResultSongInfoVO":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"$ref":"#/components/schemas/SongInfoVO"}},"title":"数据响应规范结构"},"SongInfoVO":{"type":"object","description":"歌曲详情VO","properties":{"neteaseId":{"type":"integer","format":"int64","description":"网易云id"},"songName":{"type":"string","description":"歌曲名"},"albumId":{"type":"integer","format":"int64","description":"所属专辑Id"},"albumName":{"type":"string","description":"所属专辑名"},"singer":{"type":"string","description":"歌手名"},"releaseTime":{"type":"string","format":"date","description":"发行时间"},"duration":{"type":"integer","format":"int32","description":"时长（秒）"},"coverUrl":{"type":"string","description":"封面URL"},"avgScore":{"type":"number","description":"综合评分"},"scoreCount":{"type":"integer","format":"int32","description":"评分总数"},"collectCount":{"type":"integer","format":"int32","description":"收藏总数"},"commentCount":{"type":"integer","format":"int32","description":"评论总数"}},"title":"SongInfoVO"},"ResultInteger":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"type":"integer","format":"int32"}},"title":"数据响应规范结构"},"ResultBoolean":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"type":"boolean"}},"title":"数据响应规范结构"},"AlbumInfoVO":{"type":"object","description":"专辑详情VO","properties":{"albumName":{"type":"string","description":"专辑名"},"artists":{"type":"string","description":"歌手名"},"releaseTime":{"type":"string","format":"date","description":"发行时间"},"coverUrl":{"type":"string","description":"封面"},"description":{"type":"string","description":"简介"},"avgScore":{"type":"number","description":"综合评分"},"ratingCount":{"type":"integer","format":"int32","description":"评分总数"},"collectCount":{"type":"integer","format":"int32","description":"收藏总数"},"commentCount":{"type":"integer","format":"int32","description":"评论总数"},"songs":{"type":"array","description":"专辑歌曲","items":{"$ref":"#/components/schemas/Song"}}},"title":"AlbumInfoVO"},"ResultAlbumInfoVO":{"type":"object","properties":{"状态码":{"type":"integer","format":"int32","description":"0-成功 其余-失败"},"错误信息":{"type":"string"},"响应数据":{"$ref":"#/components/schemas/AlbumInfoVO"}},"title":"数据响应规范结构"}}}}